Add files via upload
My group and I were tasked with coding a basic library system. There are two types of user for this library: admins and members.
Your system should provide the following functionality for members:
Registering with the library (a member must provide more than just a username and password to register).
2) Logging in to your library account (by providing your username and password).
3) Viewing the details of all books in the library.
4) Viewing the details of all loans currently active for you.
5) Viewing the details of all loans you have made since joining the library.
6) Borrowing a copy of a book in the library (if there are no copies available in the library, this will not succeed). A
member can only borrow one copy of any title in a single transaction. However, once they return their copy of a
book, they can borrow it again in the future.
7) Returning a book you currently have on loan.
8) Logging out.